+++ Tests/BardCore/utility/light_test.cpp
@@ -9,7 +9,7 @@ namespace testing
     TEST(light_test, constructor)
     {
         constexpr bardcore::point3d position(1, 2, 3);
-        constexpr float intensity = 100;
+        constexpr double intensity = 100;
 
         constexpr bardcore::utility::light light(position, intensity);
 
@@ -20,7 +20,7 @@ namespace testing
     TEST(light_test, copy_constructor)
     {
         constexpr bardcore::point3d position(1, 2, 3);
-        constexpr float intensity = 100;
+        constexpr double intensity = 100;
         constexpr bardcore::utility::light light(position, intensity);
 
         constexpr bardcore::utility::light light2(light);
@@ -32,7 +32,7 @@ namespace testing
     TEST(light_test, move_constructor)
     {
         constexpr bardcore::point3d position(1, 2, 3);
-        constexpr float intensity = 100;
+        constexpr double intensity = 100;
         constexpr bardcore::utility::light light(position, intensity);
 
         constexpr bardcore::utility::light light2(std::move(light));
@@ -46,29 +46,29 @@ namespace testing
         constexpr bardcore::utility::light light(bardcore::point3d(0, 0, 0), 1);
         constexpr bardcore::point3d point(1, 1, 1);
 
-        constexpr float result = light.inverse_square_law(point);
+        constexpr double result = light.inverse_square_law(point);
 
-        EXPECT_NEAR(result, 1.0f / 3.0f, ROUND_THREE_DECIMALS);
+        EXPECT_NEAR(result, 1.0 / 3.0, ROUND_THREE_DECIMALS);
     }
 
     TEST(light_test, inverse_square_law_length2)
     {
         constexpr bardcore::utility::light light(bardcore::point3d(0, 0, 0), 100);
-        constexpr float length = 25;
+        constexpr double length = 25;
 
-        constexpr float result = light.inverse_square_law(length);
+        constexpr double result = light.inverse_square_law(length);
 
-        EXPECT_NEAR(result, 100.0f / 625.0f, ROUND_THREE_DECIMALS);
+        EXPECT_NEAR(result, 100.0 / 625.0, ROUND_THREE_DECIMALS);
     }
 
     TEST(light_test, inverse_square_law_length)
     {
         constexpr bardcore::utility::light light(bardcore::point3d(0, 0, 0), 1);
-        constexpr float length = 1;
+        constexpr double length = 1;
 
-        constexpr float result = light.inverse_square_law(length);
+        constexpr double result = light.inverse_square_law(length);
 
-        EXPECT_NEAR(result, 1.0f, ROUND_THREE_DECIMALS);
+        EXPECT_NEAR(result, 1.0, ROUND_THREE_DECIMALS);
     }
 
     TEST(light_test, inverse_square_law_point2)
@@ -76,9 +76,9 @@ namespace testing
         constexpr bardcore::utility::light light(bardcore::point3d(-5, 3, 2), 100);
         constexpr bardcore::point3d point(7, 9, -25);
 
-        constexpr float result = light.inverse_square_law(point);
+        constexpr double result = light.inverse_square_law(point);
 
-        EXPECT_NEAR(result, 100.0f / 909.0f, ROUND_THREE_DECIMALS);
+        EXPECT_NEAR(result, 100.0 / 909.0, ROUND_THREE_DECIMALS);
     }
 
     //exception tests
@@ -96,8 +96,8 @@ namespace testing
     TEST(light_test, inverse_square_law_length_exception)
     {
         constexpr bardcore::utility::light light(bardcore::point3d(0, 0, 0), 1);
-        constexpr float length = 0;
-        constexpr float length2 = 10;
+        constexpr double length = 0;
+        constexpr double length2 = 10;
 
         EXPECT_THROW(light.inverse_square_law(length), bardcore::exception::zero_exception);
         EXPECT_NO_THROW(light.inverse_square_law(length2));
